Split thoth managers to remove waiting
