+++ kebechet/managers/thoth_provenance/thoth_provenance.py
@@ -49,53 +49,17 @@ class ThothProvenanceManager(ManagerBase):
         self._cached_merge_requests = None
         super().__init__(*args, **kwargs)
 
-    @property
-    def sha(self):
-        """Get SHA of the current head commit."""
-        return self.repo.head.commit.hexsha
-
-    def _issue_provenance_error(self, prov_results: list, labels: list):
-        error_info = prov_results[0]
-        text_block = ""
-
-        for err in error_info:
-            _LOGGER.info(f"{err['id']}: {err['package_name']} {err['package_version']}")
-            text_block = (
-                text_block
-                + f"## {err['type']}: {err['id']} - {err['package_name']}:{err['package_version']}\n"
-                + f"**Justification: {err['justification']}**\n"
-                + f"#### source: \n {pprint.pformat(err['source'])}\n"
-                + f"#### lock_info:\n {pprint.pformat(err['package_locked'])}"
-            )
-
-        checksum = hashlib.md5(text_block.encode("utf-8")).hexdigest()[:10]
-
-        self.sm.open_issue_if_not_exist(
-            f"{checksum} - {len(error_info)}: Automated kebechet thoth-provenance Issue",
-            lambda: text_block,
-            labels=labels,
-        )
-
     def run(self, labels: list):
         """Run the provenance check bot."""
         with cloned_repo(self.service_url, self.slug, depth=1) as repo:
             self.repo = repo
-            if os.path.isfile("Pipfile") and os.path.isfile("Pipfile.lock"):
-                res = lib.provenance_check_here()
-                for i in range(1, 11):
-                    if res is not None:
-                        break
-                    _LOGGER.warning(f"Provenance check failed, retrying ({i}/10)")
-                    res = lib.provenance_check_here(force=True)
-
-            if res is None:
-                _LOGGER.error("Provenance check failed on server side, contact the maintainer")
-                return
-
-            if res[1] is False:
-                _LOGGER.info("Provenance check found problems, creating issue...")
-                self._issue_provenance_error(res, labels)
+            if not (os.path.isfile("Pipfile") and os.path.isfile("Pipfile.lock")):
+                _LOGGER.warning("Pipfile or Pipfile.lock is missing from repo, opening issue")
+                self.sm.open_issue_if_not_exist(
+                    "Missing pipenv files",
+                    lambda: "Check your repository to make sure Pipfile and Pipfile.lock exist.",
+                    labels=labels
+                )
                 return False
-            else:
-                _LOGGER.info("Provenance check found no problems, carry on coding :)")
-                return True
+        lib.provenance_check_here(nowait=True, origin=(self.service_url + self.slug))
+        return True
